[fix] 修复test_fork问题 [style] 重整代码风格，添加注释 1. 用户程序使用print打印字符串，目前不能打印整数 2. process.c和vmem.c下新增函数注释

+++ include/kernel/process.h
@@ -58,36 +58,37 @@ typedef struct context // loongarch 12个
 #endif
 
 // Per-process state
-typedef struct proc {
-	spinlock_t lock;				///< 自旋锁限制修改
-	void *chan;						///< 如果 non-zero，sleeping on chan
-	struct proc *parent;			///< Parent process
+typedef struct proc
+{
+    spinlock_t lock;     ///< 自旋锁限制修改
+    void *chan;          ///< 如果 non-zero，sleeping on chan
+    struct proc *parent; ///< Parent process
 
-	enum procstate state; 			///< Process state
-    int  exit_state;                ///< 进程退出状态
-    int killed;                     ///< 如果不为0，则进程被杀死
-	int pid; 						///< Process ID
-	uint64 virt_addr;            	///< Virtual address of proc
-	uint64 sz;                   	///< Size of process memory (bytes)
-	uint64 kstack; 					///< Virtual address of kernel stack
-	struct trapframe *trapframe; 	///< data page for trampoline.S
-	struct context context; 		///< swtch() here to run process
-	pgtbl_t pagetable;       		///< User page table
-	int utime;						///< 用户态运行时间
-	int ktime;						///< 内核态运行时间
+    enum procstate state;        ///< Process state
+    int exit_state;              ///< 进程退出状态
+    int killed;                  ///< 如果不为0，则进程被杀死
+    int pid;                     ///< Process ID
+    uint64 virt_addr;            ///< Virtual address of proc
+    uint64 sz;                   ///< Size of process memory (bytes)
+    uint64 kstack;               ///< Virtual address of kernel stack
+    struct trapframe *trapframe; ///< data page for trampoline.S
+    struct context context;      ///< swtch() here to run process
+    pgtbl_t pagetable;           ///< User page table
+    int utime;                   ///< 用户态运行时间
+    int ktime;                   ///< 内核态运行时间
 } proc_t;
 
-struct proc*	curr_proc();
-void 			proc_init();
-void 			scheduler() __attribute__((noreturn));
-struct proc*	allocproc();
-void proc_freepagetable(struct proc * p, uint64 sz);
-void 			proc_mapstacks(pgtbl_t pagetable);
-void            sleep_on_chan(void*, struct spinlock*);
-void            wakeup(void*);
-void 			yield(void);
-uint64          fork(void);
-int             wait(int addr);
-void            exit(int exit_state);
+struct proc *curr_proc();
+void proc_init();
+void scheduler() __attribute__((noreturn));
+struct proc *allocproc();
+void proc_freepagetable(struct proc *p, uint64 sz);
+void proc_mapstacks(pgtbl_t pagetable);
+void sleep_on_chan(void *, struct spinlock *);
+void wakeup(void *);
+void yield(void);
+uint64 fork(void);
+int wait(uint64 addr);
+void exit(int exit_state);
 void reg_info(void);
 #endif // PROC_H
